+++ packages/server/skel/tailwind/app.js
@@ -1,5 +1,5 @@
 
 //
-const { app } = require('igo');
+const { app } = require('@igojs/server');
 
 app.run();