+++ packages/server/skel/tailwind/vite.config.js
@@ -1,6 +1,6 @@
 
 
 //
-const viteConfig = require('igo').dev.viteConfig;
+const viteConfig = require('@igojs/server').dev.viteConfig;
 
 module.exports = viteConfig;